DROID-3105 refactoring
