+++ core-models/src/main/java/com/anytypeio/anytype/core_models/Block.kt
@@ -31,7 +31,6 @@ data class Block(
         val featuredRelations: List<String>? by default
         val name: String? by default
         val iconEmoji: String? by default
-        val iconOption: Double? by default
         val coverId: Id? by default
         val coverType: Double? by default
         val iconImage: Id? get() = map.getSingleValue<Id>(Relations.ICON_IMAGE)
@@ -42,7 +41,6 @@ data class Block(
         val done: Boolean? by default
         val lang: String? by default
         val fileExt: String? by default
-        val fileMimeType: String? by default
         val type: List<String>
             get() = when (val value = map[TYPE_KEY]) {
                 is String -> listOf(value)
@@ -59,10 +57,6 @@ data class Block(
                 is Int -> value.toDouble()
                 else -> null
             }
-
-
-        val analyticsContext: String? by default
-        val analyticsOriginalId: String? by default
 
         companion object {
             fun empty(): Fields = Fields(emptyMap())
