+++ domain/src/main/java/com/anytypeio/anytype/domain/primitives/FieldParser.kt
@@ -1,6 +1,10 @@
 package com.anytypeio.anytype.domain.primitives
 
+import com.anytypeio.anytype.core_models.MAX_SNIPPET_SIZE
+import com.anytypeio.anytype.core_models.ObjectType
 import com.anytypeio.anytype.core_models.ObjectWrapper
+import com.anytypeio.anytype.core_models.Relations
+import com.anytypeio.anytype.core_models.SupportedLayouts
 import com.anytypeio.anytype.core_models.primitives.Field
 import com.anytypeio.anytype.core_models.primitives.FieldDateValue
 import com.anytypeio.anytype.core_models.primitives.TimestampInSeconds
@@ -8,6 +12,7 @@ import com.anytypeio.anytype.core_models.primitives.Value
 import com.anytypeio.anytype.domain.debugging.Logger
 import com.anytypeio.anytype.domain.misc.DateProvider
 import javax.inject.Inject
+import kotlin.collections.contains
 
 interface FieldParser {
     fun toDate(any: Any?): Field.Date?
@@ -26,6 +31,7 @@ class FieldParserImpl @Inject constructor(
             is Value.Single -> {
                 calculateFieldDate(value = value)
             }
+
             else -> {
                 return null
             }
@@ -33,7 +39,7 @@ class FieldParserImpl @Inject constructor(
     }
 
     override fun getObjectName(objectWrapper: ObjectWrapper.Basic): String {
-        return objectWrapper.name.orEmpty()
+        return objectWrapper.getProperObjectName().orEmpty()
     }
 
     private fun calculateFieldDate(value: Value.Single<Long>?): Field.Date? {
@@ -50,5 +56,46 @@ class FieldParserImpl @Inject constructor(
                 )
             )
         )
+    }
+
+    private fun ObjectWrapper.Basic.getProperObjectName(): String? {
+        return when (layout) {
+            ObjectType.Layout.DATE -> {
+                getProperDateName()
+            }
+
+            ObjectType.Layout.NOTE -> {
+                snippet?.replace("\n", " ")?.take(MAX_SNIPPET_SIZE)
+            }
+
+            in SupportedLayouts.fileLayouts -> {
+                val fileName = if (name.isNullOrBlank()) "Untitled" else name.orEmpty()
+                if (fileExt.isNullOrBlank()) {
+                    fileName
+                } else {
+                    if (fileName.endsWith(".$fileExt")) {
+                        fileName
+                    } else {
+                        "$fileName.$fileExt"
+                    }
+                }
+            }
+
+            else -> {
+                name
+            }
+        }
+    }
+
+    private fun ObjectWrapper.Basic.getProperDateName(): String {
+        val timestampInSeconds = getSingleValue<Double>(Relations.TIMESTAMP)?.toLong()
+        if (timestampInSeconds != null) {
+            val (formattedDate, _) = dateProvider.formatTimestampToDateAndTime(
+                timestamp = timestampInSeconds * 1000,
+            )
+            return formattedDate
+        } else {
+            return ""
+        }
     }
 }
