+++ presentation/src/main/java/com/anytypeio/anytype/presentation/editor/editor/ext/ContentTextExt.kt
@@ -3,6 +3,7 @@ package com.anytypeio.anytype.presentation.editor.editor.ext
 import com.anytypeio.anytype.core_models.Block
 import com.anytypeio.anytype.core_models.Relations
 import com.anytypeio.anytype.core_models.ext.replaceRangeWithWord
+import com.anytypeio.anytype.domain.primitives.FieldParser
 import com.anytypeio.anytype.presentation.editor.editor.Markup
 import com.anytypeio.anytype.presentation.editor.editor.Markup.Companion.NON_EXISTENT_OBJECT_MENTION_NAME
 import com.anytypeio.anytype.presentation.extension.getProperObjectName
@@ -11,7 +12,8 @@ import timber.log.Timber
 
 fun Block.Content.Text.getTextAndMarks(
     details: Block.Details,
-    marks: List<Markup.Mark>
+    marks: List<Markup.Mark>,
+    fieldParser: FieldParser
 ): Pair<String, List<Markup.Mark>> {
     if (details.details.isEmpty() ||
         marks.none { it is Markup.Mark.Mention }
