+++ presentation/src/main/java/com/anytypeio/anytype/presentation/sets/filter/FilterViewModel.kt
@@ -36,7 +36,7 @@ import com.anytypeio.anytype.presentation.extension.logEvent
 import com.anytypeio.anytype.presentation.extension.toConditionView
 import com.anytypeio.anytype.presentation.extension.type
 import com.anytypeio.anytype.presentation.mapper.toDomain
-import com.anytypeio.anytype.presentation.objects.SupportedLayouts
+import com.anytypeio.anytype.core_models.SupportedLayouts
 import com.anytypeio.anytype.presentation.objects.toCreateFilterObjectView
 import com.anytypeio.anytype.presentation.relations.FilterInputValueParser
 import com.anytypeio.anytype.presentation.relations.toCreateFilterCheckboxView
