+++ presentation/src/test/java/com/anytypeio/anytype/presentation/collections/CollectionAddRelationTest.kt
@@ -3,7 +3,6 @@ package com.anytypeio.anytype.presentation.collections
 import android.util.Log
 import app.cash.turbine.test
 import com.anytypeio.anytype.core_models.Block
-import com.anytypeio.anytype.core_models.DVViewerRelation
 import com.anytypeio.anytype.core_models.Event
 import com.anytypeio.anytype.core_models.ObjectType
 import com.anytypeio.anytype.core_models.ObjectWrapper
@@ -14,7 +13,6 @@ import com.anytypeio.anytype.core_models.Relations.ID
 import com.anytypeio.anytype.core_models.Relations.LAYOUT
 import com.anytypeio.anytype.core_models.StubRelationObject
 import com.anytypeio.anytype.presentation.objects.ObjectIcon
-import com.anytypeio.anytype.presentation.objects.getProperName
 import com.anytypeio.anytype.presentation.relations.model.DefaultObjectRelationValueView
 import com.anytypeio.anytype.presentation.sets.DataViewViewState
 import com.anytypeio.anytype.presentation.sets.ObjectSetViewModel
@@ -132,7 +130,7 @@ class CollectionAddRelationTest : ObjectSetViewModelTestSetup() {
                         items = listOf(
                             Viewer.ListView.Item.Default(
                                 objectId = objectCollection.obj1.id,
-                                name = objectCollection.obj1.getProperName(),
+                                name = objectCollection.obj1.name.orEmpty(),
                                 description = objectCollection.obj1.description,
                                 hideIcon = false,
                                 icon = ObjectIcon.Empty.Page,
@@ -153,7 +151,7 @@ class CollectionAddRelationTest : ObjectSetViewModelTestSetup() {
                             ),
                             Viewer.ListView.Item.Default(
                                 objectId = objectCollection.obj2.id,
-                                name = objectCollection.obj2.getProperName(),
+                                name = objectCollection.obj2.name.orEmpty(),
                                 description = objectCollection.obj2.description,
                                 hideIcon = false,
                                 icon = ObjectIcon.Empty.Page,
